test(compaction): run compaction scenarios against both online and offline compaction
- add compaction_scenario_test, a parameterized fixture that runs each
  compaction scenario once per mode (online / offline), so that a scenario
  added for one mode is automatically exercised by the other
- state the mode-independent expectations once: the compacted file contents,
  the compaction catalog, the snapshot rebuilt at startup and the blob files
  still referenced must be identical in both modes
- expose the two intended differences explicitly through
  retains_detached_pwals() (only online compaction keeps the rotated source
  pwals) and records_rotation_epoch() (offline compaction records the durable
  epoch read from the epoch file, which lags when an epoch carries no data)
- migrate no_pwals, scenario03, scenario04 and scenario_blob from
  compaction_test to the new fixture; scenario01 and scenario02 stay online
  only, as the lifecycle of the detached pwals has no offline counterpart
- move the tglogutil invocation helpers to compaction_test so that the
  scenario fixture and offline_compaction_test share them
