+++ test/limestone/compaction/compaction_test.cpp
@@ -23,28 +23,6 @@ using namespace limestone::api;
 using namespace limestone::internal;
     
 
-TEST_F(compaction_test, no_pwals) {
-    gen_datastore();
-    auto pwals = extract_pwal_files_from_datastore();
-    EXPECT_TRUE(pwals.empty());
-
-    compaction_catalog catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-
-    datastore_->switch_epoch(1);
-    run_compact_with_epoch_switch(2);
-
-    // No PWALs are present, so the catalog should not be updated.
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_TRUE(pwals.empty());
-}
-
 TEST_F(compaction_test, scenario01) {
 
     gen_datastore();
@@ -619,540 +597,6 @@ TEST_F(compaction_test, scenario02) {
     ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 4);
     ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 2);
     ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 1);
-}
-
-// This test case verifies the correct behavior of `remove_entry`.
-TEST_F(compaction_test, scenario03) {
-
-    // 1. Create multiple PWALs using two different storage IDs
-    gen_datastore();
-    datastore_->switch_epoch(1);
-
-    // Storage ID 1: key1 added, then removed
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key1", "value1", {1, 0});
-    lc0_->remove_entry(1, "key1", {1, 1});  // use remove_entry
-    lc0_->end_session();
-
-    // Storage ID 2: key2 added, no removal
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key2", "value2", {1, 0});
-    lc1_->end_session();
-
-    // Storage ID 1: key3 removed first, then added
-    lc2_->begin_session();
-    lc2_->remove_entry(1, "key3", {1, 0});
-    lc2_->add_entry(1, "key3", "value3", {1, 3});
-    lc2_->end_session();
-
-    // Storeage ID 1: key4 deleted witout adding
-    lc0_->begin_session();
-    lc0_->remove_entry(1, "key4", {1, 0});
-    lc0_->end_session();
-
-    datastore_->switch_epoch(2);
-
-    // Check the created PWAL files
-    auto pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 3);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000");
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0001");
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0002");
-
-    auto log_entries = read_log_file("pwal_0000", location);
-    ASSERT_EQ(log_entries.size(), 3);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 1, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key1", std::nullopt, 1, 1, {}, log_entry::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key4", std::nullopt, 1, 0, {}, log_entry::entry_type::remove_entry));
-    log_entries = read_log_file("pwal_0001", location);
-    ASSERT_EQ(log_entries.size(), 1);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 2, "key2", "value2", 1, 0, {}, log_entry::entry_type::normal_entry));
-    log_entries = read_log_file("pwal_0002", location);
-    ASSERT_EQ(log_entries.size(), 2);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key3", std::nullopt, 1, 0, {}, log_entry ::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key3", "value3", 1, 3, {}, log_entry::entry_type::normal_entry));
-
-    // 2. Execute compaction
-    run_compact_with_epoch_switch(3);
-
-    // Check the catalog and PWALs after compaction
-    compaction_catalog catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 2);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 1);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 3);
-
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 4);  // Includes the compacted file
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 2);                                                                                 // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key3", "value3", 0, 0, {}, log_entry::entry_type::normal_entry));  // write version changed to 0
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 2, "key2", "value2", 0, 0, {}, log_entry::entry_type::normal_entry));  // write version changed to 0
-
-    // 3. Add/Update PWALs (include remove_entry again)
-
-    // Storage ID 1: key1 added, then removed
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key11", "value1", {2, 0});
-    lc0_->remove_entry(1, "key11", {2, 1});  // use remove_entry
-    lc0_->end_session();
-
-    // Storage ID 2: key2 added, no removal
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key21", "value2", {2, 0});
-    lc1_->end_session();
-
-    // Storage ID 1: key3 removed first, then added
-    lc2_->begin_session();
-    lc2_->remove_entry(1, "key31", {2, 0});
-    lc2_->add_entry(1, "key31", "value3", {2, 3});
-    lc2_->end_session();
-
-    // Storeage ID 1: key4 deleted witout adding
-    lc0_->begin_session();
-    lc0_->remove_entry(1, "key41", {2, 0});
-    lc0_->end_session();
-
-    datastore_->switch_epoch(4);
-    pwals = extract_pwal_files_from_datastore();
-
-    // Check the created PWAL files
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 7);  // 3 new pwals and 3 rotaed pwals and 1 compacted file
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000");
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0001");
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0002");
-
-    log_entries = read_log_file("pwal_0000", location);
-    ASSERT_EQ(log_entries.size(), 3);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key11", "value1", 2, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key11", std::nullopt, 2, 1, {}, log_entry::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key41", std::nullopt, 2, 0, {}, log_entry::entry_type::remove_entry));
-    log_entries = read_log_file("pwal_0001", location);
-    ASSERT_EQ(log_entries.size(), 1);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 2, "key21", "value2", 2, 0, {}, log_entry::entry_type::normal_entry));
-    log_entries = read_log_file("pwal_0002", location);
-    ASSERT_EQ(log_entries.size(), 2);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key31", std::nullopt, 2, 0, {}, log_entry::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key31", "value3", 2, 3, {}, log_entry::entry_type::normal_entry));
-
-    // 4. Restart the datastore
-    datastore_->shutdown();
-    datastore_ = nullptr;
-    gen_datastore();  // Restart
-
-    // 5. check the compacted file and snapshot creating at the boot time
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 2);                                                                                 // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key3", "value3", 0, 0, {}, log_entry::entry_type::normal_entry));  // write version changed to 0
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 2, "key2", "value2", 0, 0, {}, log_entry::entry_type::normal_entry));  // write version changed to 0
-
-    log_entries = read_log_file("data/snapshot", location);
-    ASSERT_EQ(log_entries.size(), 4);  // Ensure that there are log entries
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key11", std::nullopt, 2, 1, {}, log_entry::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key31", "value3", 2, 3, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key41", std::nullopt, 2, 0, {}, log_entry::entry_type::remove_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 2, "key21", "value2", 2, 0, {}, log_entry::entry_type::normal_entry));
-
-    // 5. Verify the snapshot contents after restart
-    std::vector<std::pair<std::string, std::string>> kv_list = restart_datastore_and_read_snapshot();
-
-    // key1 should exist with its updated value, key2 and key3 should be removed
-    ASSERT_EQ(kv_list.size(), 4);
-    EXPECT_EQ(kv_list[0].first, "key3");
-    EXPECT_EQ(kv_list[0].second, "value3");
-    EXPECT_EQ(kv_list[1].first, "key31");
-    EXPECT_EQ(kv_list[1].second, "value3");
-    EXPECT_EQ(kv_list[2].first, "key2");
-    EXPECT_EQ(kv_list[2].second, "value2");
-    EXPECT_EQ(kv_list[3].first, "key21");
-    EXPECT_EQ(kv_list[3].second, "value2");
-}
-
-// This test case verifies the correct behavior of `remove_storage`.
-TEST_F(compaction_test, scenario04) {
-
-    gen_datastore();
-    datastore_->switch_epoch(1);
-
-    // Storage ID 1: Add normal entries
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key1", "value1", {1, 0});
-    lc0_->add_entry(1, "key2", "value2", {1, 1});
-    lc0_->end_session();
-
-    // Storage ID 2: Add normal entries
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key3", "value3", {1, 0});
-    lc1_->add_entry(2, "key4", "value4", {1, 1});
-    lc1_->end_session();
-
-    // Storage ID 1: Add more normal entries
-    lc2_->begin_session();
-    lc2_->add_entry(1, "key5", "value5", {1, 2});
-    lc2_->add_entry(1, "key6", "value6", {1, 3});
-    lc2_->end_session();
-
-    // Advance the epoch to 2
-    datastore_->switch_epoch(2);
-
-    // Remove storage for Storage ID 2
-    lc1_->begin_session();
-    lc1_->remove_storage(2, {2, 0});  // Removes the storage with ID 2
-    lc1_->end_session();
-
-    // Advance the epoch to 3
-    datastore_->switch_epoch(3);
-
-    // Add an entry to Storage ID 1
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key7", "value7", {3, 0});
-    lc0_->end_session();
-
-    // Add an entry to Storage ID 2
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key8", "value8", {3, 0});
-    lc1_->end_session();
-
-    // Chek PWALs before compaction
-    auto pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 3);
-
-    std::vector<log_entry> log_entries = read_log_file("pwal_0000", location);
-    ASSERT_EQ(log_entries.size(), 3);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 1, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key2", "value2", 1, 1, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key7", "value7", 3, 0, {}, log_entry::entry_type::normal_entry));
-
-    log_entries = read_log_file("pwal_0001", location);
-    ASSERT_EQ(log_entries.size(), 4);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 2, "key3", "value3", 1, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 2, "key4", "value4", 1, 1, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 2, "", "", 2, 0, {}, log_entry::entry_type::remove_storage));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 2, "key8", "value8", 3, 0, {}, log_entry::entry_type::normal_entry));
-
-    log_entries = read_log_file("pwal_0002", location);
-    ASSERT_EQ(log_entries.size(), 2);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key5", "value5", 1, 2, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key6", "value6", 1, 3, {}, log_entry::entry_type::normal_entry));
-
-    // online compaction
-    run_compact_with_epoch_switch(4);
-
-    // Check PWALs after compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 4);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 1);
-
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 6);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key2", "value2", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key5", "value5", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 1, "key6", "value6", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[4], 1, "key7", "value7", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[5], 2, "key8", "value8", 0, 0, {}, log_entry::entry_type::normal_entry));
-
-    // Storage ID 1: Add normal entries
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key11", "value1", {4, 0});
-    lc0_->add_entry(1, "key12", "value2", {4, 1});
-    lc0_->end_session();
-
-    // Storage ID 2: Add normal entries
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key13", "value3", {4, 0});
-    lc1_->add_entry(2, "key14", "value4", {4, 1});
-    lc1_->end_session();
-
-    // Storage ID 1: Add more normal entries
-    lc2_->begin_session();
-    lc2_->add_entry(1, "key15", "value5", {4, 2});
-    lc2_->add_entry(1, "key16", "value6", {4, 3});
-    lc2_->end_session();
-
-    // Advance the epoch to 5
-    datastore_->switch_epoch(5);
-
-    // Remove storage for Storage ID 1
-    lc1_->begin_session();
-    lc1_->remove_storage(1, {5, 0});  // Removes the storage with ID 2
-    lc1_->end_session();
-
-    // Advance the epoch to 6
-    datastore_->switch_epoch(6);
-
-    // Add an entry to Storage ID 1
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key17", "value7", {6, 0});
-    lc0_->end_session();
-
-    // Add an entry to Storage ID 2
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key18", "value8", {6, 0});
-    lc1_->end_session();
-
-    // Advance the epoch to 6
-    datastore_->switch_epoch(7);
-
-    // Chek newly created PWALs
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 7);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 1);
-
-    log_entries = read_log_file("pwal_0000", location);
-    ASSERT_EQ(log_entries.size(), 3);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key11", "value1", 4, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key12", "value2", 4, 1, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key17", "value7", 6, 0, {}, log_entry::entry_type::normal_entry));
-
-    log_entries = read_log_file("pwal_0001", location);
-    ASSERT_EQ(log_entries.size(), 4);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 2, "key13", "value3", 4, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 2, "key14", "value4", 4, 1, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "", "", 5, 0, {}, log_entry::entry_type::remove_storage));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 2, "key18", "value8", 6, 0, {}, log_entry::entry_type::normal_entry));
-
-    log_entries = read_log_file("pwal_0002", location);
-    ASSERT_EQ(log_entries.size(), 2);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key15", "value5", 4, 2, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key16", "value6", 4, 3, {}, log_entry::entry_type::normal_entry));
-
-    // Restart the datastore
-
-    std::vector<std::pair<std::string, std::string>> kv_list = restart_datastore_and_read_snapshot();
-
-    // check the compacted file and snapshot creating at the boot time
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 6);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key2", "value2", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key5", "value5", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 1, "key6", "value6", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[4], 1, "key7", "value7", 0, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[5], 2, "key8", "value8", 0, 0, {}, log_entry::entry_type::normal_entry));
-
-    log_entries = read_log_file("data/snapshot", location);
-    ASSERT_EQ(log_entries.size(), 4);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key17", "value7", 6, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 2, "key13", "value3", 4, 0, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 2, "key14", "value4", 4, 1, {}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 2, "key18", "value8", 6, 0, {}, log_entry::entry_type::normal_entry));
-
-    // 5. Verify the snapshot contents after restart
-
-    // key1 should exist with its updated value, key2 and key3 should be removed
-    ASSERT_EQ(kv_list.size(), 5);
-    EXPECT_EQ(kv_list[0].first, "key17");
-    EXPECT_EQ(kv_list[0].second, "value7");
-    EXPECT_EQ(kv_list[1].first, "key13");
-    EXPECT_EQ(kv_list[1].second, "value3");
-    EXPECT_EQ(kv_list[2].first, "key14");
-    EXPECT_EQ(kv_list[2].second, "value4");
-    EXPECT_EQ(kv_list[3].first, "key18");
-    EXPECT_EQ(kv_list[3].second, "value8");
-    EXPECT_EQ(kv_list[4].first, "key8");
-    EXPECT_EQ(kv_list[4].second, "value8");
-}
-
-// This test case verifies the correct behavior of blob feature.
-TEST_F(compaction_test, scenario_blob) {
-
-    gen_datastore();
-    datastore_->switch_epoch(1);
-
-    compaction_catalog catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_max_blob_id(), 0);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-    EXPECT_EQ(datastore_->next_blob_id(), 1);
-
-    // No PWALs are present => The catalog should not be updated.
-    run_compact_with_epoch_switch(2);
-
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_max_blob_id(), 0);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-
-
-    // Update the max_blob_id in the catalog
-    catalog.update_catalog_file(0, 123, {}, {});
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_max_blob_id(), 123);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-
-    // No PWALs are present => The max_blob_id in the catalog should not be updated.
-    run_compact_with_epoch_switch(3);
-
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 0);
-    EXPECT_EQ(catalog.get_max_blob_id(), 123);
-    EXPECT_EQ(catalog.get_compacted_files().size(), 0);
-    EXPECT_EQ(catalog.get_detached_pwals().size(), 0);
-
-    // Create two PWALs containing BLOBs
-
-    lc0_->begin_session();
-    lc0_->add_entry(1, "key1", "value1", {1, 0}, {1001, 1002});
-    lc0_->add_entry(1, "key2", "value2", {1, 1}, {1003, 1004});
-    lc0_->end_session();
-
-    // Storage ID 2: Add normal entries
-    lc1_->begin_session();
-    lc1_->add_entry(2, "key3", "value3", {1, 0}, {1005, 1006});
-    lc1_->end_session();
-
-    datastore_->switch_epoch(4);
-
-    // Check PWALs before compaction
-    auto pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 2);
-
-    std::vector<log_entry> log_entries = read_log_file("pwal_0000", location);
-    ASSERT_EQ(log_entries.size(), 2);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 1, 0, {1001, 1002}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key2", "value2", 1, 1, {1003, 1004}, log_entry::entry_type::normal_with_blob));
-
-    log_entries = read_log_file("pwal_0001", location);
-    ASSERT_EQ(log_entries.size(), 1);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 2, "key3", "value3", 1, 0, {1005, 1006}, log_entry::entry_type::normal_with_blob));
-
-    // online compaction
-    run_compact_with_epoch_switch(5);
-
-    // Check compaction catalog
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 4);
-    EXPECT_EQ(catalog.get_max_blob_id(), 1006);
-    ASSERT_EQ(catalog.get_compacted_files().size(), 1);
-    EXPECT_TRUE(catalog.get_compacted_files().find(compacted_file_info("pwal_0000.compacted", 1)) != catalog.get_compacted_files().end());
-    ASSERT_EQ(catalog.get_detached_pwals().size(), 2);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[0].find("pwal_0000.") == 0);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[1].find("pwal_0001.") == 0);
-
-    // Check PWALs after compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 3);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 3);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 0, 0, {1001, 1002}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key2", "value2", 0, 0, {1003, 1004}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 2, "key3", "value3", 0, 0, {1005, 1006}, log_entry::entry_type::normal_with_blob));
-
-    // Write entries without BLOBs and perform compaction.
-
-    // Storage ID 1: Add more normal entries
-    lc2_->begin_session();
-    lc2_->add_entry(1, "key15", "value5", {4, 2});
-    lc2_->add_entry(1, "key16", "value6", {4, 3});
-    lc2_->end_session();
-
-    // Check PWALs before compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 4);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0002");
-
-    // online compaction
-    run_compact_with_epoch_switch(6);
-
-    // Check compaction catalog
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 5);
-    EXPECT_EQ(catalog.get_max_blob_id(), 1006);
-    ASSERT_EQ(catalog.get_compacted_files().size(), 1);
-    EXPECT_TRUE(catalog.get_compacted_files().find(compacted_file_info("pwal_0000.compacted", 1)) != catalog.get_compacted_files().end());
-    ASSERT_EQ(catalog.get_detached_pwals().size(), 3);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[0].find("pwal_0000.") == 0);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[1].find("pwal_0001.") == 0);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[2].find("pwal_0002.") == 0);
-
-    // Check PWALs after compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 4);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 1);
-
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 5);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 0, 0, {1001, 1002}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key15", "value5", 0, 0, {1001, 1002}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key16", "value6", 0, 0, {1001, 1002}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 1, "key2", "value2", 0, 0, {1003, 1004}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[4], 2, "key3", "value3", 0, 0, {1005, 1006}, log_entry::entry_type::normal_with_blob));
-
-    // Write entries with BLOBs but with blob_ids smaller than max_blob_id and perform compaction.
-
-    // Storage ID 1: Add more normal entries
-    lc2_->begin_session();
-    lc2_->add_entry(2, "key5", "value5", {4, 2}, {128, 32, 59});
-    lc2_->end_session();
-
-    // Check PWALs before compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 5);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 1);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0002");
-
-    // online compaction
-    run_compact_with_epoch_switch(7);
-
-    // Check compaction catalog
-    catalog = compaction_catalog::from_catalog_file(location);
-    EXPECT_EQ(catalog.get_max_epoch_id(), 6);
-    EXPECT_EQ(catalog.get_max_blob_id(), 1006);
-    ASSERT_EQ(catalog.get_compacted_files().size(), 1);
-    EXPECT_TRUE(catalog.get_compacted_files().find(compacted_file_info("pwal_0000.compacted", 1)) != catalog.get_compacted_files().end());
-    ASSERT_EQ(catalog.get_detached_pwals().size(), 4);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[0].find("pwal_0000.") == 0);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[1].find("pwal_0001.") == 0);
-    EXPECT_TRUE(get_sorted_list(catalog.get_detached_pwals())[2].find("pwal_0002.") == 0);
-
-    // Check PWALs after compaction
-    pwals = extract_pwal_files_from_datastore();
-    EXPECT_EQ(pwals.size(), 5);
-    ASSERT_PRED_FORMAT2(ContainsString, pwals, "pwal_0000.compacted");
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0000.", 2);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0001.", 1);
-    ASSERT_PRED_FORMAT3(ContainsPrefix, pwals, "pwal_0002.", 2);
-
-    log_entries = read_log_file("pwal_0000.compacted", location);
-    ASSERT_EQ(log_entries.size(), 6);
-    EXPECT_TRUE(AssertLogEntry(log_entries[0], 1, "key1", "value1", 0, 0, {1001, 1002}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[1], 1, "key15", "value5", 0, 0, {1001, 1002}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[2], 1, "key16", "value6", 0, 0, {1001, 1002}, log_entry::entry_type::normal_entry));
-    EXPECT_TRUE(AssertLogEntry(log_entries[3], 1, "key2", "value2", 0, 0, {1003, 1004}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[4], 2, "key3", "value3", 0, 0, {1005, 1006}, log_entry::entry_type::normal_with_blob));
-    EXPECT_TRUE(AssertLogEntry(log_entries[5], 2, "key5", "value5", 0, 0, {128, 32, 59}, log_entry::entry_type::normal_with_blob));
-
-    // datastoreの再起動後、コンパクションカタログのmax_blobg_idがnext_blob_idに反映されることを確認
-    datastore_->shutdown();
-    datastore_ = nullptr;
-    gen_datastore();
-    EXPECT_EQ(datastore_->next_blob_id(), 1007);
 }
 
 // This test is disabled because it is environment-dependent and may not work properly in CI environments.
